Fixed missing code for 'first start' skip button
As you fire up the Application for the first time, it shows a form that tries to be as much user friendly as possible, to teach a first-timer how to use the application.
For other users, this can be skipped with a button or just by exiting the form:
the problem was that I forgot (nice, Blasted.... Nice) to put any form of code into the button pressing event.