New GRAPHICS! (Also a couple of functional and useful tools)
I've been in this days working on adding some essential features to the App...
I'VE ADDED VOLUME CONTROL FOR VIDEOS... Finally
I've made media files smaller than the desktop resolution centre the screen (though for an esthetic idea, they will be anchored to the top)
Now the only thing missing is a Zoom Control to resize the media file on the desktop.
Have a nice weekend.