Ready to Merge & Release + Farewell Youtube
I've been working in this hours on different things. Also I've made a little trailer that I've uploaded on Youtube (https://youtu.be/e8PXMGN8YT8)
And I've tested the Online Content feature a couple of times:
I've tweaked the UI, removing the laggy image preview inside the application
And then... I had to remove youtube link support, as the webcomponent that I use is not compatible with the Youtube iframe API, so it's a GOODBYE (for now).
Cya